+++ crates/futures/src/lib.rs
@@ -32,6 +32,7 @@
 
 #![cfg_attr(target_feature = "atomics", feature(stdarch_wasm_atomic_wait))]
 #![deny(missing_docs)]
+#![cfg_attr(docsrs, feature(doc_cfg))]
 
 use js_sys::Promise;
 use std::cell::RefCell;
@@ -43,6 +44,7 @@ use std::task::{Context, Poll, Waker};
 use wasm_bindgen::prelude::*;
 
 mod queue;
+#[cfg_attr(docsrs, doc(cfg(feature = "futures-core-03-stream")))]
 #[cfg(feature = "futures-core-03-stream")]
 pub mod stream;