+++ crates/web-sys/src/features/gen_Headers.rs
@@ -49,6 +49,20 @@ extern "C" {
     #[doc = ""]
     #[doc = "*This API requires the following crate features to be activated: `Headers`*"]
     pub fn delete(this: &Headers, name: &str) -> Result<(), JsValue>;
+    # [wasm_bindgen (method , structural , js_class = "Headers" , js_name = entries)]
+    #[doc = "The `entries()` method."]
+    #[doc = ""]
+    #[doc = "[MDN Documentation](https://developer.mozilla.org/en-US/docs/Web/API/Headers/entries)"]
+    #[doc = ""]
+    #[doc = "*This API requires the following crate features to be activated: `Headers`*"]
+    pub fn entries(this: &Headers) -> ::js_sys::Iterator;
+    # [wasm_bindgen (catch , method , structural , js_class = "Headers" , js_name = forEach)]
+    #[doc = "The `forEach()` method."]
+    #[doc = ""]
+    #[doc = "[MDN Documentation](https://developer.mozilla.org/en-US/docs/Web/API/Headers/forEach)"]
+    #[doc = ""]
+    #[doc = "*This API requires the following crate features to be activated: `Headers`*"]
+    pub fn for_each(this: &Headers, callback: &::js_sys::Function) -> Result<(), JsValue>;
     # [wasm_bindgen (catch , method , structural , js_class = "Headers" , js_name = get)]
     #[doc = "The `get()` method."]
     #[doc = ""]
@@ -63,6 +77,13 @@ extern "C" {
     #[doc = ""]
     #[doc = "*This API requires the following crate features to be activated: `Headers`*"]
     pub fn has(this: &Headers, name: &str) -> Result<bool, JsValue>;
+    # [wasm_bindgen (method , structural , js_class = "Headers" , js_name = keys)]
+    #[doc = "The `keys()` method."]
+    #[doc = ""]
+    #[doc = "[MDN Documentation](https://developer.mozilla.org/en-US/docs/Web/API/Headers/keys)"]
+    #[doc = ""]
+    #[doc = "*This API requires the following crate features to be activated: `Headers`*"]
+    pub fn keys(this: &Headers) -> ::js_sys::Iterator;
     # [wasm_bindgen (catch , method , structural , js_class = "Headers" , js_name = set)]
     #[doc = "The `set()` method."]
     #[doc = ""]
@@ -70,4 +91,11 @@ extern "C" {
     #[doc = ""]
     #[doc = "*This API requires the following crate features to be activated: `Headers`*"]
     pub fn set(this: &Headers, name: &str, value: &str) -> Result<(), JsValue>;
+    # [wasm_bindgen (method , structural , js_class = "Headers" , js_name = values)]
+    #[doc = "The `values()` method."]
+    #[doc = ""]
+    #[doc = "[MDN Documentation](https://developer.mozilla.org/en-US/docs/Web/API/Headers/values)"]
+    #[doc = ""]
+    #[doc = "*This API requires the following crate features to be activated: `Headers`*"]
+    pub fn values(this: &Headers) -> ::js_sys::Iterator;
 }